requirements.txt
boto3==1.18.10
google-cloud-storage==1.41.1
pytest==6.2.5
